pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools-scm >= 8"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "async-kafka"
requires-python = ">=3.10"
authors = [{ name = "Timofei Kukushkin", email = "tima@kukushkin.me" }]
description = "Thin async wrapper for Confluent Kafka Python client"
readme = "README.md"
dependencies = [
  "anyio",
  "confluent-kafka",
]
license = { file = "LICENSE" }
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Framework :: AsyncIO",
  "Framework :: Trio",
]

[dependency-groups]
test = [
  "pytest",
  "pytest-anyio",
  "pytest-cov",
  "lovely-pytest-docker",
  "trio",
]
lint = [
  { include-group = "test" },
  "pyright",
  "ruff",
]
docs = [
  "furo",
  "sphinx",
  "sphinx-autodoc-typehints",
  "myst-parser",
]
dev = [
  { include-group = "test" },
  { include-group = "lint" },
  { include-group = "docs" },
  "pytest-sugar",
  "tox",
  "tox-uv",
  "uv",
]

[project.urls]
Homepage = "https://github.com/tkukushkin/async-kafka"
Repository = "https://github.com/tkukushkin/async-kafka.git"
Issues = "https://github.com/tkukushkin/async-kafka/issues"
Documentation = "https://async-kafka.readthedocs.io/stable"
Changelog = "https://github.com/tkukushkin/async-kafka/releases"

[tool.setuptools_scm]
version_file = "src/async_kafka/_version.py"
